Various simulation updates (#1500)
- Compile the Quartus simulation library for VCS in the background during
  gen_sim_files.sh. The pre-compiled images are now copied into both unit test
  and UVM builds.
- Unit tests now use the partial compilation flow in VCS.
- Unify the unit test build rules for VCS and VCS-MX.
- Limit waveform dumping in UVM to the DUT to reduce output.

+++ sim/unit_test/scripts/sim_setup_common.sh
@@ -39,7 +39,6 @@ else
 fi
 echo "entering sim_setup_common.sh: SIM_DIR: $SIM_DIR"
 
-IP_LIB_DIR=$TEST_DIR
 IP_SIM_SCRIPT_DIR="$OFS_ROOTDIR/sim/scripts/qip_sim_script"
 
 # ----------------------------------------
@@ -52,32 +51,23 @@ rm -rf $SIM_DIR
 # ----------------------------------------
 mkdir $SIM_DIR
 
-if [ $VCSMX -eq 1 ]; then
-   # IP library compilation
-   if [ $SKIP_IP_CMP -eq 0 ]; then
-      if [ -d "$IP_LIB_DIR/ip_libraries" ]; then
-         rm -rf "$IP_LIB_DIR/ip_libraries" 
-      fi 
-      
-      mkdir -p $IP_LIB_DIR/ip_libraries
-      cp -f $IP_SIM_SCRIPT_DIR/synopsys/vcsmx/synopsys_sim.setup $IP_LIB_DIR/ip_libraries
-      cd $IP_LIB_DIR/ip_libraries && $IP_SIM_SCRIPT_DIR/synopsys/vcsmx/vcsmx_setup.sh SKIP_SIM=1 QSYS_SIMDIR=$IP_SIM_SCRIPT_DIR QUARTUS_INSTALL_DIR=$QUARTUS_ROOTDIR USER_DEFINED_ELAB_OPTIONS="-xlrm\ uniq_prior_final"
-   fi
-
-   # Simulation setup
-   echo WORK \> DEFAULT > $SIM_DIR/synopsys_sim.setup
-   echo DEFAULT \: worklib >>  $SIM_DIR/synopsys_sim.setup              
-   mkdir  $SIM_DIR/worklib
-   rsync -avz --checksum --ignore-times ${IP_LIB_DIR}/ip_libraries/* $SIM_DIR
-fi
-
 if [ $MSIM -eq 1 ]; then
    cp ${IP_SIM_SCRIPT_DIR}/../msim_filelist.sh $SIM_DIR 
 else
+   # Pre-compiled Quartus libraries
+   if [ ! -f "${IP_SIM_SCRIPT_DIR}"/quartus_libs/vcsmx/synopsys_sim.setup ]; then
+      echo ""
+      echo "Quartus simulation library not found in "${IP_SIM_SCRIPT_DIR}"/quartus_libs/vcsmx/synopsys_sim.setup"
+      echo "This library is supposed to be pre-compiled by gen_sim_files.sh."
+      echo "Aborting..."
+      exit 1
+   fi
+   rsync -a "${IP_SIM_SCRIPT_DIR}"/quartus_libs/vcsmx/ "${SIM_DIR}"/
+
    cp ${IP_SIM_SCRIPT_DIR}/../vcs_filelist.sh $SIM_DIR 
 fi
-cp ${IP_SIM_SCRIPT_DIR}/../rtl_pcie.f $SIM_DIR
 
+cp ${IP_SIM_SCRIPT_DIR}/../rtl_pcie.f $SIM_DIR
 
 if [ $MSIM -eq 1 ]; then
    sed -i 's/PCIE_RTL_FILELIST=.*/PCIE_RTL_FILELIST="-f .\/rtl_pcie.f"/' ${SIM_DIR}/msim_filelist.sh
